+++ src/rockverse/configure/__init__.py
@@ -63,9 +63,6 @@ class Config():
         else:
             raise ValueError('NO')
 
-
-
-
     @property
     def mpi_rank(self):
         """
@@ -100,42 +97,6 @@ class Config():
         Returns the list of all GPUs available to the processor in ``processor_name``.
         """
         return self._gpus
-
-    @property
-    def selected_gpus(self):
-        """Get or set the list of selected GPU indices.
-
-        These indices set the devices that are allowed to be used by
-        RockVerse during runtime.
-
-        You can set this list at runtime by providing a list of integers in
-        range('total number of available GPUs').
-
-        Examples
-        --------
-            >>> # Get the currently selected GPU indices
-            >>> current_selected = config.selected_gpus
-
-            >>> # Set the selected GPUs to the first and second GPUs available
-            >>> config['selected_gpus'] = [0, 1]
-
-            >>> # You can use any iterable
-            >>> config['selected_gpus'] = (0, 1, 2)
-            >>> config['selected_gpus'] = {0, 1, 2}
-            >>> config['selected_gpus'] = range(2)
-
-
-            >>> # Attempting to set selected GPUs to an invalid index will raise an error
-            >>> try:
-            >>>     config['selected_gpus'] = [0, 5]  # Assuming only 3 GPUs are available
-            >>> except RuntimeError as e:
-            >>>    print(e)  # Output: GPU device indices must be less than 3.
-
-            >>> # Setting selected GPUs to an empty list means no GPU will be used
-            >>> config['selected_gpus'] = []
-        """
-        return self['selected_gpus']
-
 
 
     def print_available_gpus(self):
@@ -197,6 +158,27 @@ config = Config()
 
 
 class config_context():
+    """
+    Context manager for temporarily modifying the configuration settings.
+
+    This context manager allows for a temporary reassignment of configuration parameters.
+    Changes made within the context manager are reverted back to the original settings once
+    the context is exited.
+
+    Parameters:
+    -----------
+    kwargs : dict
+        A dictionary of configuration parameters to update.
+
+    Example:
+    --------
+
+    Temporarily change the list of allowed GPU devices:
+
+    >>> with config_context({'selected_gpus': [2, 4, 6]}):
+    >>>     # Do stuff using these 3 devices
+    >>> # After the with block, the original configuration is restored.
+    """
 
     def __init__(self, kwargs):
         self.backup_dict = {}