+++ Changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ``dualenergyct`` module renamed to ``dect``.
 - ``rc`` module merged into ``configure`` module.
+- Added ``config_context`` context manager for runtime configurations.
 
 ## [1.0.0] - 2025-03-10